fix compiler warnings, seen with Qt 5.15, that apply to older Qt versions too
These changes are in master since quite long and unconditionally, see
b7b35643a1.

+++ mscore/editdrumset.cpp
@@ -69,7 +69,7 @@ NoteHead::Group noteHeadNames[] = {
 bool EditDrumsetTreeWidgetItem::operator<(const QTreeWidgetItem & other) const
       {
       if (treeWidget()->sortColumn() == Column::PITCH)
-            return data(Column::PITCH, Qt::UserRole) < other.data(Column::PITCH, Qt::UserRole);
+            return data(Column::PITCH, Qt::UserRole).toInt() < other.data(Column::PITCH, Qt::UserRole).toInt();
       else
             return QTreeWidgetItem::operator<(other);
       }
